# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: contact@josehernandez.dev
    name: Smidgen
    url: https://example.com/
  description: API for interacting with Smidgen.
  license:
    name: " GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007"
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: Smidgen
  version: 0.0.1
servers:
- url: /api/v1
paths:
  /user/:
    get:
      description: Get all Users stored in the database.
      operationId: get_user_user__get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/user'
                title: Response Get User User  Get
                type: array
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
      summary: Get Users
      tags:
      - user
    post:
      description: Create a new user in the database.
      operationId: add_user_user__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create user
      tags:
      - user
  /user/{user_id}:
    delete:
      description: Delete the specified user from the database.
      operationId: delete_user_user__user_id__delete
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
        style: simple
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete user
      tags:
      - user
    get:
      description: Get the specified user stored in the database.
      operationId: get_User_user__user_id__get
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get user
      tags:
      - user
    put:
      description: Update the specified user in the database.
      operationId: update_user_user__user_id__put
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update user
      tags:
      - user
...
components:
  schemas:
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      example:
        msg: msg
        loc:
        - Location_inner
        - Location_inner
        type: type
      properties:
        loc:
          items:
            $ref: '#/components/schemas/Location_inner'
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    user:
      example:
        business_unit_id: 6
        user_id: 0
        email: email
        username: username
      properties:
        user_id:
          title: User Id
          type: integer
        business_unit_id:
          title: Business Unit Id
          type: integer
        username:
          title: Username
          type: string
        email:
          title: Email
          type: string
      required:
      - business_unit_id
      - email
      - user_id
      - username
      title: user
      type: object
    Location_inner:
      anyOf:
      - type: string
      - type: integer